@@ -3,22 +3,16 @@ using Core.Entities;
 using Core.Interfaces;
 using Infra.Kafka.Interfaces;
 
-namespace Application.Services
+namespace Application.Services;
+
+public class PersonEventService(IKafkaProducer kafkaProducer) : IPersonEventService
 {
-	public class PersonEventService : IPersonEventService
+	private readonly IKafkaProducer _kafkaProducer = kafkaProducer;
+
+	public async Task ProduceEvent(string topicName, Person person)
 	{
-		private readonly IKafkaProducer _kafkaProducer;
+		var messageBody = JsonSerializer.Serialize(person);
 
-		public PersonEventService(IKafkaProducer kafkaProducer)
-		{
-			_kafkaProducer = kafkaProducer;
-		}
-
-		public async Task ProduceEvent(string topicName, Person person)
-		{
-			var messageBody = JsonSerializer.Serialize(person);
-
-			await _kafkaProducer.ProduceEvent(topicName, messageBody);
-		}
+		await _kafkaProducer.ProduceEvent(topicName, messageBody);
 	}
 }
