@@ -1,8 +1,7 @@
-namespace Core.Entities
-{
-	public sealed record Person(
-		Guid id,
-		string name,
-		int age
-	);
-}
+namespace Core.Entities;
+
+public sealed record Person(
+	Guid Id,
+	string Name,
+	int Age
+);
